qna_generator/qna_generator/requirements.txt
streamlit
requests
beautifulsoup4
pymupdf
python-docx
openai


